Backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0.0
transformers
av==15.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
dotenv==0.9.9
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
packaging==25.0
protobuf==6.32.0
pycparser==2.22
pyreadline3==3.5.4
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
sounddevice==0.5.2
sympy==1.13.1
tokenizers==0.22.0
torch==2.5.1+cu121
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
webrtcvad-wheels==2.0.14
